Dramatically changed the underlying way that the snake is rendered
Previously, each render call the program would look to each segment's parent (the one before it)
and its child, (the one after it) to determine what type of sprite should be used for that segment.
This resulted in a lot of unnecessary computation and an increase in code complexity.

Now, instead of the graphical data of each segment being computed each time the snake is drawn,
computation only occurs when necessary. Essentially, the code now requires a segment to only copy the data of its parent segment
after updating, in a very similar manner to how segment positions are calculated in order to move the rest of the snake body along.